day2 tasks updated

+++ src/day2/Task9.java
@@ -1,19 +1,17 @@
 package day2;
 
-public class Task9
-{
-    public static void main(String[] args)
-    {
+public class Task9 {
+    public static void main(String[] args) {
         String kelime = "abc";
 
         char ilkHarf = kelime.charAt(0);
-        int ilkHarfASCII = (int) ilkHarf;
+        int ilkHarfASCII = ilkHarf;
 
         char ikinciHarf = kelime.charAt(1);
-        int ikinciHarfASCII = (int) ikinciHarf;
+        int ikinciHarfASCII = ikinciHarf;
 
         char ucuncuHarf = kelime.charAt(2);
-        int ucuncuHarfASCII = (int) ucuncuHarf;
+        int ucuncuHarfASCII = ucuncuHarf;
 
         System.out.println("Karakter: " + ilkHarf + ", ASCII Değeri: " + ilkHarfASCII);
         System.out.println("Karakter: " + ikinciHarf + ", ASCII Değeri: " + ikinciHarfASCII);
